Extract view mixins into separate files
