+++ packages/ember-views/lib/views/view.js
@@ -3,49 +3,21 @@
 // Ember.ContainerView circular dependency
 // Ember.ENV
 import Ember from 'ember-metal/core';
-import create from 'ember-metal/platform/create';
 
 import Evented from "ember-runtime/mixins/evented";
 import EmberObject from "ember-runtime/system/object";
 import EmberError from "ember-metal/error";
 import { get } from "ember-metal/property_get";
-import { set } from "ember-metal/property_set";
-import setProperties from "ember-metal/set_properties";
 import run from "ember-metal/run_loop";
 import { addObserver, removeObserver } from "ember-metal/observer";
-import { defineProperty } from "ember-metal/properties";
-import {
-  guidFor,
-  typeOf
-} from "ember-metal/utils";
+import { guidFor } from "ember-metal/utils";
 import { computed } from "ember-metal/computed";
 import {
   Mixin,
   observer,
   beforeObserver
 } from "ember-metal/mixin";
-import KeyStream from "ember-views/streams/key_stream";
-import StreamBinding from "ember-metal/streams/stream_binding";
-import ContextStream from "ember-views/streams/context_stream";
-import { read } from "ember-metal/streams/utils";
-
-import AttrNode from "ember-views/attr_nodes/attr_node";
-
-import { typeOf } from "ember-metal/utils";
 import { deprecateProperty } from "ember-metal/deprecate_property";
-import { A as emberA } from "ember-runtime/system/native_array";
-
-import {
-  streamifyClassNameBinding
-} from "ember-views/streams/class_name_binding";
-
-// ES6TODO: functions on EnumerableUtils should get their own export
-import {
-  forEach,
-  addObject,
-  removeObject
-} from "ember-metal/enumerable_utils";
-
 import {
   propertyWillChange,
   propertyDidChange
@@ -55,50 +27,21 @@ import jQuery from "ember-views/system/jquery";
 import "ember-views/system/ext";  // for the side effect of extending Ember.run.queues
 
 import CoreView from "ember-views/views/core_view";
-import {
-  subscribe,
-  read,
-  isStream
-} from "ember-metal/streams/utils";
-
-import { canSetNameOnInputs } from "ember-views/system/platform";
+import ViewStreamSupport from "ember-views/mixins/view_stream_support";
+import ViewKeywordSupport from "ember-views/mixins/view_keyword_support";
+import ViewContextSupport from "ember-views/mixins/view_context_support";
+import { default as ViewChildViewsSupport, childViewsProperty } from "ember-views/mixins/view_child_views_support";
+import ViewStateSupport from "ember-views/mixins/view_state_support";
+import TemplateRenderingSupport from "ember-views/mixins/template_rendering_support";
+import ClassNamesSupport from "ember-views/mixins/class_names_support";
+import AttributeBindingsSupport from "ember-views/mixins/attribute_bindings_support";
 
 function K() { return this; }
-
-// Circular dep
-var _renderView;
-function renderView(view, buffer, template) {
-  if (_renderView === undefined) {
-    _renderView = require('ember-htmlbars/system/render-view')['default'];
-  }
-  _renderView(view, buffer, template);
-}
 
 /**
 @module ember
 @submodule ember-views
 */
-var childViewsProperty = computed(function() {
-  var childViews = this._childViews;
-  var ret = emberA();
-
-  forEach(childViews, function(view) {
-    var currentChildViews;
-    if (view.isVirtual) {
-      if (currentChildViews = get(view, 'childViews')) {
-        ret.pushObjects(currentChildViews);
-      }
-    } else {
-      ret.push(view);
-    }
-  });
-
-  ret.replace = function (idx, removedCount, addedViews) {
-    throw new EmberError("childViews is immutable");
-  };
-
-  return ret;
-});
 
 Ember.warn("The VIEW_PRESERVES_CONTEXT flag has been removed and the functionality can no longer be disabled.", Ember.ENV.VIEW_PRESERVES_CONTEXT !== false);
 
@@ -114,616 +57,6 @@ Ember.warn("The VIEW_PRESERVES_CONTEXT flag has been removed and the functionali
 Ember.TEMPLATES = {};
 
 var EMPTY_ARRAY = [];
-
-var ViewStreamSupport = Mixin.create({
-  init: function() {
-    this._baseContext = undefined;
-    this._contextStream = undefined;
-    this._streamBindings = undefined;
-    this._super.apply(this, arguments);
-  },
-
-  getStream: function(path) {
-    var stream = this._getContextStream().get(path);
-
-    stream._label = path;
-
-    return stream;
-  },
-
-  _willDestroyElement: function() {
-    if (this._streamBindings) {
-      this._destroyStreamBindings();
-    }
-    if (this._contextStream) {
-      this._destroyContextStream();
-    }
-  },
-
-  _getBindingForStream: function(pathOrStream) {
-    if (this._streamBindings === undefined) {
-      this._streamBindings = create(null);
-    }
-
-    var path = pathOrStream;
-    if (isStream(pathOrStream)) {
-      path = pathOrStream._label;
-
-      if (!path) {
-        // if no _label is present on the provided stream
-        // it is likely a subexpr and cannot be set (so it
-        // does not need a StreamBinding)
-        return pathOrStream;
-      }
-    }
-
-    if (this._streamBindings[path] !== undefined) {
-      return this._streamBindings[path];
-    } else {
-      var stream = this._getContextStream().get(path);
-      var streamBinding = new StreamBinding(stream);
-
-      streamBinding._label = path;
-
-      return this._streamBindings[path] = streamBinding;
-    }
-  },
-
-  _destroyStreamBindings: function() {
-    var streamBindings = this._streamBindings;
-    for (var path in streamBindings) {
-      streamBindings[path].destroy();
-    }
-    this._streamBindings = undefined;
-  },
-
-  _getContextStream: function() {
-    if (this._contextStream === undefined) {
-      this._baseContext = new KeyStream(this, 'context');
-      this._contextStream = new ContextStream(this);
-    }
-
-    return this._contextStream;
-  },
-
-  _destroyContextStream: function() {
-    this._baseContext.destroy();
-    this._baseContext = undefined;
-    this._contextStream.destroy();
-    this._contextStream = undefined;
-  },
-
-  _unsubscribeFromStreamBindings: function() {
-    for (var key in this._streamBindingSubscriptions) {
-      var streamBinding = this[key + 'Binding'];
-      var callback = this._streamBindingSubscriptions[key];
-      streamBinding.unsubscribe(callback);
-    }
-  }
-});
-
-var ViewKeywordSupport = Mixin.create({
-  init: function() {
-    this._super.apply(this, arguments);
-
-    if (!this._keywords) {
-      this._keywords = create(null);
-    }
-    this._keywords._view = this;
-    this._keywords.view = undefined;
-    this._keywords.controller = new KeyStream(this, 'controller');
-    this._setupKeywords();
-  },
-
-  _setupKeywords: function() {
-    var keywords = this._keywords;
-    var contextView = this._contextView || this._parentView;
-
-    if (contextView) {
-      var parentKeywords = contextView._keywords;
-
-      keywords.view = this.isVirtual ? parentKeywords.view : this;
-
-      for (var name in parentKeywords) {
-        if (keywords[name]) {
-          continue;
-        }
-
-        keywords[name] = parentKeywords[name];
-      }
-    } else {
-      keywords.view = this.isVirtual ? null : this;
-    }
-  }
-});
-
-var ViewContextSupport = Mixin.create({
-  /**
-    The object from which templates should access properties.
-
-    This object will be passed to the template function each time the render
-    method is called, but it is up to the individual function to decide what
-    to do with it.
-
-    By default, this will be the view's controller.
-
-    @property context
-    @type Object
-  */
-  context: computed(function(key, value) {
-    if (arguments.length === 2) {
-      set(this, '_context', value);
-      return value;
-    } else {
-      return get(this, '_context');
-    }
-  }).volatile(),
-
-  /**
-    Private copy of the view's template context. This can be set directly
-    by Handlebars without triggering the observer that causes the view
-    to be re-rendered.
-
-    The context of a view is looked up as follows:
-
-    1. Supplied context (usually by Handlebars)
-    2. Specified controller
-    3. `parentView`'s context (for a child of a ContainerView)
-
-    The code in Handlebars that overrides the `_context` property first
-    checks to see whether the view has a specified controller. This is
-    something of a hack and should be revisited.
-
-    @property _context
-    @private
-  */
-  _context: computed(function(key, value) {
-    if (arguments.length === 2) {
-      return value;
-    }
-
-    var parentView, controller;
-
-    if (controller = get(this, 'controller')) {
-      return controller;
-    }
-
-    parentView = this._parentView;
-    if (parentView) {
-      return get(parentView, '_context');
-    }
-
-    return null;
-  }),
-
-  _controller: null,
-
-  /**
-    The controller managing this view. If this property is set, it will be
-    made available for use by the template.
-
-    @property controller
-    @type Object
-  */
-  controller: computed(function(key, value) {
-    if (arguments.length === 2) {
-      this._controller = value;
-      return value;
-    }
-
-    if (this._controller) {
-      return this._controller;
-    }
-
-    var parentView = this._parentView;
-    return parentView ? get(parentView, 'controller') : null;
-  })
-});
-
-var ViewChildViewsSupport = Mixin.create({
-  /**
-    Array of child views. You should never edit this array directly.
-    Instead, use `appendChild` and `removeFromParent`.
-
-    @property childViews
-    @type Array
-    @default []
-    @private
-  */
-  childViews: childViewsProperty,
-
-  _childViews: EMPTY_ARRAY,
-
-  init: function() {
-    // setup child views. be sure to clone the child views array first
-    this._childViews = this._childViews.slice();
-
-    this._super.apply(this, arguments);
-  },
-
-  appendChild: function(view, options) {
-    return this.currentState.appendChild(this, view, options);
-  },
-
-  /**
-    Removes the child view from the parent view.
-
-    @method removeChild
-    @param {Ember.View} view
-    @return {Ember.View} receiver
-  */
-  removeChild: function(view) {
-    // If we're destroying, the entire subtree will be
-    // freed, and the DOM will be handled separately,
-    // so no need to mess with childViews.
-    if (this.isDestroying) { return; }
-
-    // update parent node
-    set(view, '_parentView', null);
-
-    // remove view from childViews array.
-    var childViews = this._childViews;
-
-    removeObject(childViews, view);
-
-    this.propertyDidChange('childViews'); // HUH?! what happened to will change?
-
-    return this;
-  },
-
-  /**
-    Instantiates a view to be added to the childViews array during view
-    initialization. You generally will not call this method directly unless
-    you are overriding `createChildViews()`. Note that this method will
-    automatically configure the correct settings on the new view instance to
-    act as a child of the parent.
-
-    @method createChildView
-    @param {Class|String} viewClass
-    @param {Hash} [attrs] Attributes to add
-    @return {Ember.View} new instance
-  */
-  createChildView: function(maybeViewClass, _attrs) {
-    if (!maybeViewClass) {
-      throw new TypeError("createChildViews first argument must exist");
-    }
-
-    if (maybeViewClass.isView && maybeViewClass._parentView === this && maybeViewClass.container === this.container) {
-      return maybeViewClass;
-    }
-
-    var attrs = _attrs || {};
-    var view;
-    attrs._parentView = this;
-    attrs.renderer = this.renderer;
-
-    if (maybeViewClass.isViewClass) {
-      attrs.container = this.container;
-
-      view = maybeViewClass.create(attrs);
-
-      // don't set the property on a virtual view, as they are invisible to
-      // consumers of the view API
-      if (view.viewName) {
-        set(get(this, 'concreteView'), view.viewName, view);
-      }
-    } else if ('string' === typeof maybeViewClass) {
-      var fullName = 'view:' + maybeViewClass;
-      var ViewKlass = this.container.lookupFactory(fullName);
-
-      Ember.assert("Could not find view: '" + fullName + "'", !!ViewKlass);
-
-      view = ViewKlass.create(attrs);
-    } else {
-      view = maybeViewClass;
-      Ember.assert('You must pass instance or subclass of View', view.isView);
-
-      attrs.container = this.container;
-      setProperties(view, attrs);
-    }
-
-    return view;
-  }
-});
-
-var ViewStateSupport = Mixin.create({
-  transitionTo: function(state, children) {
-    Ember.deprecate("Ember.View#transitionTo has been deprecated, it is for internal use only");
-    this._transitionTo(state, children);
-  },
-
-  _transitionTo: function(state, children) {
-    var priorState = this.currentState;
-    var currentState = this.currentState = this._states[state];
-    this._state = state;
-
-    if (priorState && priorState.exit) { priorState.exit(this); }
-    if (currentState.enter) { currentState.enter(this); }
-  }
-});
-
-var TemplateRenderingSupport = Mixin.create({
-  /**
-    Called on your view when it should push strings of HTML into a
-    `Ember.RenderBuffer`. Most users will want to override the `template`
-    or `templateName` properties instead of this method.
-
-    By default, `Ember.View` will look for a function in the `template`
-    property and invoke it with the value of `context`. The value of
-    `context` will be the view's controller unless you override it.
-
-    @method render
-    @param {Ember.RenderBuffer} buffer The render buffer
-  */
-  render: function(buffer) {
-    // If this view has a layout, it is the responsibility of the
-    // the layout to render the view's template. Otherwise, render the template
-    // directly.
-    var template = get(this, 'layout') || get(this, 'template');
-    renderView(this, buffer, template);
-  }
-});
-
-var ClassNamesSupport = Mixin.create({
-  concatenatedProperties: ['classNames', 'classNameBindings'],
-
-  init: function() {
-    this._super.apply(this, arguments);
-
-    Ember.assert("Only arrays are allowed for 'classNameBindings'", typeOf(this.classNameBindings) === 'array');
-    this.classNameBindings = emberA(this.classNameBindings.slice());
-
-    Ember.assert("Only arrays of static class strings are allowed for 'classNames'. For dynamic classes, use 'classNameBindings'.", typeOf(this.classNames) === 'array');
-    this.classNames = emberA(this.classNames.slice());
-  },
-
-  /**
-    Standard CSS class names to apply to the view's outer element. This
-    property automatically inherits any class names defined by the view's
-    superclasses as well.
-
-    @property classNames
-    @type Array
-    @default ['ember-view']
-  */
-  classNames: ['ember-view'],
-
-  /**
-    A list of properties of the view to apply as class names. If the property
-    is a string value, the value of that string will be applied as a class
-    name.
-
-    ```javascript
-    // Applies the 'high' class to the view element
-    Ember.View.extend({
-      classNameBindings: ['priority']
-      priority: 'high'
-    });
-    ```
-
-    If the value of the property is a Boolean, the name of that property is
-    added as a dasherized class name.
-
-    ```javascript
-    // Applies the 'is-urgent' class to the view element
-    Ember.View.extend({
-      classNameBindings: ['isUrgent']
-      isUrgent: true
-    });
-    ```
-
-    If you would prefer to use a custom value instead of the dasherized
-    property name, you can pass a binding like this:
-
-    ```javascript
-    // Applies the 'urgent' class to the view element
-    Ember.View.extend({
-      classNameBindings: ['isUrgent:urgent']
-      isUrgent: true
-    });
-    ```
-
-    This list of properties is inherited from the view's superclasses as well.
-
-    @property classNameBindings
-    @type Array
-    @default []
-  */
-  classNameBindings: EMPTY_ARRAY,
-
-  /**
-    Iterates over the view's `classNameBindings` array, inserts the value
-    of the specified property into the `classNames` array, then creates an
-    observer to update the view's element if the bound property ever changes
-    in the future.
-
-    @method _applyClassNameBindings
-    @private
-  */
-  _applyClassNameBindings: function() {
-    var classBindings = this.classNameBindings;
-
-    if (!classBindings || !classBindings.length) { return; }
-
-    var classNames = this.classNames;
-    var elem, newClass, dasherizedClass;
-
-    // Loop through all of the configured bindings. These will be either
-    // property names ('isUrgent') or property paths relative to the view
-    // ('content.isUrgent')
-    forEach(classBindings, function(binding) {
-
-      var boundBinding;
-      if (isStream(binding)) {
-        boundBinding = binding;
-      } else {
-        boundBinding = streamifyClassNameBinding(this, binding, '_view.');
-      }
-
-      // Variable in which the old class value is saved. The observer function
-      // closes over this variable, so it knows which string to remove when
-      // the property changes.
-      var oldClass;
-
-      // Set up an observer on the context. If the property changes, toggle the
-      // class name.
-      var observer = this._wrapAsScheduled(function() {
-        // Get the current value of the property
-        elem = this.$();
-        newClass = read(boundBinding);
-
-        // If we had previously added a class to the element, remove it.
-        if (oldClass) {
-          elem.removeClass(oldClass);
-          // Also remove from classNames so that if the view gets rerendered,
-          // the class doesn't get added back to the DOM.
-          classNames.removeObject(oldClass);
-        }
-
-        // If necessary, add a new class. Make sure we keep track of it so
-        // it can be removed in the future.
-        if (newClass) {
-          elem.addClass(newClass);
-          oldClass = newClass;
-        } else {
-          oldClass = null;
-        }
-      });
-
-      // Get the class name for the property at its current value
-      dasherizedClass = read(boundBinding);
-
-      if (dasherizedClass) {
-        // Ensure that it gets into the classNames array
-        // so it is displayed when we render.
-        addObject(classNames, dasherizedClass);
-
-        // Save a reference to the class name so we can remove it
-        // if the observer fires. Remember that this variable has
-        // been closed over by the observer.
-        oldClass = dasherizedClass;
-      }
-
-      subscribe(boundBinding, observer, this);
-      // Remove className so when the view is rerendered,
-      // the className is added based on binding reevaluation
-      this.one('willClearRender', function() {
-        if (oldClass) {
-          classNames.removeObject(oldClass);
-          oldClass = null;
-        }
-      });
-
-    }, this);
-  }
-});
-
-var AttributeBindingsSupport = Ember.Mixin.create({
-  concatenatedProperties: ['attributeBindings'],
-
-  /**
-    A list of properties of the view to apply as attributes. If the property is
-    a string value, the value of that string will be applied as the attribute.
-
-    ```javascript
-    // Applies the type attribute to the element
-    // with the value "button", like <div type="button">
-    Ember.View.extend({
-      attributeBindings: ['type'],
-      type: 'button'
-    });
-    ```
-
-    If the value of the property is a Boolean, the name of that property is
-    added as an attribute.
-
-    ```javascript
-    // Renders something like <div enabled="enabled">
-    Ember.View.extend({
-      attributeBindings: ['enabled'],
-      enabled: true
-    });
-    ```
-
-    @property attributeBindings
-  */
-  attributeBindings: EMPTY_ARRAY,
-
-  _unspecifiedAttributeBindings: null,
-
-  /**
-    Iterates through the view's attribute bindings, sets up observers for each,
-    then applies the current value of the attributes to the passed render buffer.
-
-    @method _applyAttributeBindings
-    @param {Ember.RenderBuffer} buffer
-    @param {Array} attributeBindings
-    @private
-  */
-  _applyAttributeBindings: function(buffer) {
-    var attributeBindings = this.attributeBindings;
-
-    if (!attributeBindings || !attributeBindings.length) { return; }
-
-    var unspecifiedAttributeBindings = this._unspecifiedAttributeBindings = this._unspecifiedAttributeBindings || {};
-
-    var binding, colonIndex, property, attrName, attrNode, attrValue;
-    var i, l;
-    for (i=0, l=attributeBindings.length; i<l; i++) {
-      binding = attributeBindings[i];
-      colonIndex = binding.indexOf(':');
-      if (colonIndex === -1) {
-        property = binding;
-        attrName = binding;
-      } else {
-        property = binding.substring(0, colonIndex);
-        attrName = binding.substring(colonIndex + 1);
-      }
-
-      Ember.assert('You cannot use class as an attributeBinding, use classNameBindings instead.', attrName !== 'class');
-
-      if (property in this) {
-        attrValue = this.getStream('view.'+property);
-        attrNode = new AttrNode(attrName, attrValue);
-        this.appendAttr(attrNode);
-        if (!canSetNameOnInputs && attrName === 'name') {
-          buffer.attr('name', read(attrValue));
-        }
-      } else {
-        unspecifiedAttributeBindings[property] = attrName;
-      }
-    }
-
-    // Lazily setup setUnknownProperty after attributeBindings are initially applied
-    this.setUnknownProperty = this._setUnknownProperty;
-  },
-
-  /**
-    We're using setUnknownProperty as a hook to setup attributeBinding observers for
-    properties that aren't defined on a view at initialization time.
-
-    Note: setUnknownProperty will only be called once for each property.
-
-    @method setUnknownProperty
-    @param key
-    @param value
-    @private
-  */
-  setUnknownProperty: null, // Gets defined after initialization by _applyAttributeBindings
-
-  _setUnknownProperty: function(key, value) {
-    var attrName = this._unspecifiedAttributeBindings && this._unspecifiedAttributeBindings[key];
-
-    defineProperty(this, key);
-
-    if (attrName) {
-      var attrValue = this.getStream('view.'+key);
-      var attrNode = new AttrNode(attrName, attrValue);
-      this.appendAttr(attrNode);
-    }
-    return set(this, key, value);
-  }
-});
 
 /**
   `Ember.View` is the class in Ember responsible for encapsulating templates of
@@ -1474,7 +807,6 @@ var View = CoreView.extend(ViewStreamSupport, ViewKeywordSupport, ViewContextSup
     @default null
   */
   isVisible: true,
-
 
   // When it's a virtual view, we need to notify the parent that their
   // childViews will change.
